feat: rename app observability commands to list
