+++ shortcuts/apps/apps_logs.go
