+++ shortcuts/apps/apps_logs_test.go
