+++ shortcuts/apps/apps_traces_test.go
