+++ shortcuts/apps/shortcuts.go
@@ -26,8 +26,8 @@ func Shortcuts() []common.Shortcut {
 		withExtraTips(AppsLogGet, "Tip: logs are online-only; keep --environment omitted or set --environment online."),
 		withExtraTips(AppsTraceList, "Tip: traces are online-only; keep --environment omitted or set --environment online."),
 		withExtraTips(AppsTraceGet, "Tip: traces are online-only; keep --environment omitted or set --environment online."),
-		withExtraTips(AppsMetricQuery, "Tip: metrics are online-only; keep --environment omitted or set --environment online."),
-		withExtraTips(AppsAnalyticsQuery, "Tip: analytics are online-only; keep --environment omitted or set --environment online."),
+		withExtraTips(AppsMetricList, "Tip: metrics are online-only; keep --environment omitted or set --environment online."),
+		withExtraTips(AppsAnalyticsList, "Tip: analytics are online-only; keep --environment omitted or set --environment online."),
 		AppsEnvVarList,
 		envSet,
 		envDelete,
