+++ shortcuts/apps/shortcuts_test.go
@@ -11,7 +11,7 @@ import (
 
 // 钉死域内 shortcut 数量。少一条（漏挂）或多一条（误加）都会被这个测试拦截。
 // 6 基础 + 1 init + 3 publish + 1 env-pull
-//   - 6 observability（log-list/log-get/trace-list/trace-get/metric-query/analytics-query）
+//   - 6 observability（log-list/log-get/trace-list/trace-get/metric-list/analytics-list）
 //   - 3 env（list/set/delete）
 //   - 16 db（table-list/table-schema/sql/dev-init/data-import/data-export/changelog-list/
 //     audit-status/audit-enable/audit-disable/audit-list/
@@ -31,6 +31,15 @@ func TestAppsShortcuts_DoesNotIncludeEnvGet(t *testing.T) {
 	for _, sc := range Shortcuts() {
 		switch sc.Command {
 		case "+env-get", "+envvar-get", "+envvar-list", "+envvar-set", "+envvar-delete":
+			t.Fatalf("Shortcuts() must not register %s", sc.Command)
+		}
+	}
+}
+
+func TestAppsShortcuts_DoesNotIncludeMetricQueryAliases(t *testing.T) {
+	for _, sc := range Shortcuts() {
+		switch sc.Command {
+		case "+metric-query", "+analytics-query":
 			t.Fatalf("Shortcuts() must not register %s", sc.Command)
 		}
 	}
